Rename the ChannelListHeader to the GroupChannelListHeader
