+++ src/modules/ChannelList/components/ChannelListHeader/index.tsx
@@ -1,3 +1,3 @@
-import ChannelListHeader from '../../../GroupChannelList/components/ChannelListHeader';
+import ChannelListHeader from '../../../GroupChannelList/components/GroupChannelListHeader';
 
 export default ChannelListHeader;
