+++ src/modules/GroupChannelList/components/GroupChannelListHeader/index.scss
